requirement.txt
streamlit
azure-ai-openai
azure-core
#pip install -r requirements.txt
